// src/api/travel-warnings-service/models/UserTripDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * travel-warnings-service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UserTripDto
 */
export interface UserTripDto {
    /**
     * 
     * @type {number}
     * @memberof UserTripDto
     */
    readonly id?: number;
    /**
     * 
     * @type {string}
     * @memberof UserTripDto
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof UserTripDto
     */
    countryCode?: string;
    /**
     * 
     * @type {string}
     * @memberof UserTripDto
     */
    countryName?: string;
    /**
     * 
     * @type {Date}
     * @memberof UserTripDto
     */
    startDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof UserTripDto
     */
    endDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof UserTripDto
     */
    tripName?: string;
    /**
     * 
     * @type {boolean}
     * @memberof UserTripDto
     */
    notificationsEnabled?: boolean;
}

/**
 * Check if a given object implements the UserTripDto interface.
 */
export function instanceOfUserTripDto(value: object): value is UserTripDto {
    return true;
}

export function UserTripDtoFromJSON(json: any): UserTripDto {
    return UserTripDtoFromJSONTyped(json, false);
}

export function UserTripDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserTripDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'email': json['email'] == null ? undefined : json['email'],
        'countryCode': json['countryCode'] == null ? undefined : json['countryCode'],
        'countryName': json['countryName'] == null ? undefined : json['countryName'],
        'startDate': json['startDate'] == null ? undefined : (new Date(json['startDate'])),
        'endDate': json['endDate'] == null ? undefined : (new Date(json['endDate'])),
        'tripName': json['tripName'] == null ? undefined : json['tripName'],
        'notificationsEnabled': json['notificationsEnabled'] == null ? undefined : json['notificationsEnabled'],
    };
}

export function UserTripDtoToJSON(json: any): UserTripDto {
    return UserTripDtoToJSONTyped(json, false);
}

export function UserTripDtoToJSONTyped(value?: Omit<UserTripDto, 'id'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'email': value['email'],
        'countryCode': value['countryCode'],
        'countryName': value['countryName'],
        'startDate': value['startDate'] == null ? undefined : ((value['startDate']).toISOString().substring(0,10)),
        'endDate': value['endDate'] == null ? undefined : ((value['endDate']).toISOString().substring(0,10)),
        'tripName': value['tripName'],
        'notificationsEnabled': value['notificationsEnabled'],
    };
}